Restructure to prevent errors, one object store.
Fixes errors when cloning due to invoking git commands before the repo
was initiaized.

It also now stores everything in one object store. There really was no
good reason to create them based on the domain.

Also adds a `overwrite-alternates` to use in case of emergencies. This
really shouldn't be needed but I'm not good with picking out patterns
with sed to pick-out previous entries.